deprecate HasCrossing and SubsystemClockCrossing
Move subsystem.SubsystemClockCrossing => diplomacy.ClockCrossingType
Put abstract CrossingHelper base class in diplomacy
Make concrete AXI4CrossingHelper, TLCrossingHelper, IntCrossingHelper
Replace CrossingWrapper with AXI4CrossingWrapper, TLCrossingWrapper
